requirements.txt
discord.py
python-dotenv>=1.0.0
colorama>=0.4.6
aiohttp>=3.9.0
aiohttp-socks>=0.8.0
-e ./CatLink
PyNaCl